+++ TexTransCore/BlendTexture/ShaderAsset/MaskShader.shader
@@ -41,7 +41,7 @@ Shader "Hidden/MaskShader"
                 return o;
             }
 
-            fixed4 frag (v2f i) : SV_Target
+            float4 frag (v2f i) : SV_Target
             {
                 float4 col = tex2Dlod(_MainTex ,float4(i.uv,0,0));
                 float MaskColor =  tex2Dlod(_MaskTex ,float4(i.uv,0,0)).a;
